ProjectFiles/requirements.txt
streamlit
transformers
torch
sentencepiece
python-dotenv
gTTS
